FDG-9287 Create reusable structure
