--- src/components/published-reports/month-picker/month-picker-dropdown/month-picker-dropdown.module.scss
@@ -1,128 +1,0 @@
-@import '../../../../variables.module.scss';
-
-.dropdownContainer {
-  width: 20rem;
-  background-color: white;
-  z-index: 5;
-  position: absolute;
-  box-shadow: 0 0.25rem 1rem 0 rgba(0, 0, 0, 0.15);
-  padding-top: 1rem;
-  margin-top: 0.5rem;
-
-  .publishedDateLabel {
-    font-size: $font-size-14;
-    margin-bottom: 0.5rem;
-  }
-
-  .selectedDateDisplay {
-    border: 1px solid $light-grey-border;
-    border-radius: 3px;
-    display: flex;
-    align-items: center;
-    padding: .5rem 1rem;
-  }
-
-  .publishedDateLabel, .selectedDateDisplay {
-    margin-left: 1.5rem;
-    margin-right: 1.5rem;
-  }
-
-  .yearButton {
-    font-size: $font-size-16;
-    background: none;
-    border: none;
-    color: $font-body-copy;
-    font-weight: $semi-bold-weight;
-    padding: 0.625rem 1rem 0.625rem 0;
-    margin-top: 1rem;
-    width: 100%;
-
-    .arrowIcon {
-      padding-left: 0.5rem;
-    }
-  }
-
-  .dropdownList {
-    overflow-y: auto;
-    margin-top: -1.5rem;
-
-    ul {
-      display: flex;
-      align-items: center;
-      flex-direction: column;
-      padding: 0;
-      margin: 0;
-      height: 12.375rem;
-    }
-
-    li {
-      list-style: none;
-      display: flex;
-      justify-content: center;
-      width: 100%;
-
-      button {
-        font-size: $font-size-16;
-        background: none;
-        border: none;
-        color: $font-body-copy;
-        padding: .5rem .75rem;
-        width: 100%;
-        border-radius: 3px;
-
-        &.selected {
-          background-color: #EFF6FB;
-          color: #0071BC;
-        }
-      }
-    }
-  }
-
-  .buttonContainer {
-    padding: 1rem;
-    display: flex;
-    justify-content: flex-end;
-    border-top: 1px solid #D9D9D9;
-
-    .applyButton {
-      background-color: #0071BC;
-      border-radius: 3px;
-      border: none;
-      color: white;
-      margin-left: 1rem;
-      display: flex;
-      align-items: center;
-
-      .checkIcon {
-        padding-right: .5rem;
-        color: white;
-      }
-    }
-
-    .cancelButton {
-      text-decoration: underline;
-      background: none;
-      border: none;
-      color: $font-body-copy;
-    }
-
-    .applyButton, .cancelButton {
-      padding: .5rem 1rem;
-      font-weight: $semi-bold-weight;
-      font-size: $font-size-16;
-      cursor: pointer;
-
-    }
-
-  }
-}
-
-@media screen and (max-width: $breakpoint-sm) {
-  .dropdownContainer {
-    .yearButton {
-      padding-right: 0;
-    }
-  }
-}
-
-
