+++ src/components/published-reports/report-date-dropdown/report-date-dropdown.module.scss
@@ -1,0 +1,63 @@
+@import '../../../variables.module.scss';
+
+
+.dropdownContainer {
+  width: 20rem;
+  background-color: white;
+  z-index: 5;
+  position: absolute;
+  box-shadow: 0 0.25rem 1rem 0 rgba(0, 0, 0, 0.15);
+  padding-top: 1rem;
+  margin-top: 0.5rem;
+
+  .publishedDateLabel {
+    font-size: $font-size-14;
+    margin-bottom: 0.5rem;
+  }
+
+  .selectedDateDisplay {
+    border: 1px solid $light-grey-border;
+    border-radius: 3px;
+    display: flex;
+    align-items: center;
+    padding: .5rem 1rem;
+  }
+}
+
+.buttonContainer {
+  padding: 1rem;
+  display: flex;
+  justify-content: flex-end;
+  border-top: 1px solid #D9D9D9;
+
+  .applyButton {
+    background-color: #0071BC;
+    border-radius: 3px;
+    border: none;
+    color: white;
+    margin-left: 1rem;
+    display: flex;
+    align-items: center;
+
+    .checkIcon {
+      padding-right: .5rem;
+      color: white;
+    }
+  }
+
+  .cancelButton {
+    text-decoration: underline;
+    background: none;
+    border: none;
+    color: $font-body-copy;
+  }
+
+  .applyButton, .cancelButton {
+    padding: .5rem 1rem;
+    font-weight: $semi-bold-weight;
+    font-size: $font-size-16;
+    cursor: pointer;
+
+  }
+
+}
