+++ src/components/published-reports/reports-section/reports-section.tsx
@@ -4,7 +4,7 @@ import DownloadReportTable from '../download-report-table/download-report-table'
 import { reportsTip, note, publishDate } from './reports-section.module.scss';
 import DatasetSectionContainer from '../../dataset-section-container/dataset-section-container';
 import { getPublishedDates } from '../../../helpers/dataset-detail/report-helpers';
-import MonthPicker from '../month-picker/month-picker';
+import ReportDatePicker from '../report-date-picker/report-date-picker';
 
 export const title = 'Reports and Files';
 export interface IReports {
@@ -76,8 +76,7 @@ const ReportsSection: FunctionComponent<{ publishedReportsProp: IReports[]; data
   return (
     <div style={{ display: getDisplayStatus(publishedReportsProp) }}>
       <DatasetSectionContainer title={title} id="reports-and-files">
-        {!isDailyReport && <MonthPicker />}
-        {isDailyReport && <div className={publishDate}>Published Date:</div>}
+        <ReportDatePicker isDailyReport={isDailyReport} />
         <DownloadReportTable reports={currentReports} isDailyReport={isDailyReport} />
         {dataset?.publishedReportsTip && (
           <div className={reportsTip}>
